// .temp/packages/types/generated.ts
export type AvatarUrls = {
    xs: string;
    sm: string;
    base: string;
    lg: string;
    xl: string;
    xxl: string;
};
export type User = {
    id: string;
    avatar_url: string;
    avatar_urls: AvatarUrls;
    cover_photo_url: string | null;
    email: string;
    username: string;
    display_name: string;
    system: boolean;
    integration: boolean;
    notifications_paused: boolean;
    notification_pause_expires_at: string | null;
    timezone: string | null;
    logged_in: boolean;
    type_name: string;
};
export type OrganizationMembershipStatus = {
    message: string;
    emoji: string;
    expiration_setting: '30m' | '1h' | '4h' | 'today' | 'this_week' | 'custom';
    expires_at: string | null;
    pause_notifications: boolean;
    expires_in: '30m' | '1h' | '4h' | 'today' | 'this_week' | 'custom';
};
export type OrganizationMember = {
    id: string;
    role: 'admin' | 'member' | 'viewer' | 'guest';
    created_at: string;
    deactivated: boolean;
    is_organization_member: boolean;
    user: User;
    status: OrganizationMembershipStatus | null;
};
export type ImageUrls = {
    original_url: string;
    thumbnail_url: string;
    feed_url: string;
    email_url: string;
    slack_url: string;
    large_url: string;
};
export type Attachment = {
    id: string;
    file_type: string;
    url: string;
    app_url: string;
    download_url: string;
    preview_url: string | null;
    preview_thumbnail_url: string | null;
    image_urls: ImageUrls | null;
    link: boolean;
    image: boolean;
    video: boolean;
    audio: boolean;
    origami: boolean;
    principle: boolean;
    lottie: boolean;
    stitch: boolean;
    gif: boolean;
    duration: number;
    width: number;
    height: number;
    subject_type: string | null;
    name: string | null;
    size: number | null;
    remote_figma_url: string | null;
    no_video_track: boolean;
    gallery_id: string | null;
    type_name: string;
    subject_id: string | null;
    is_subject_comment: boolean;
    relative_url: string;
    preview_relative_url: string | null;
    comments_count: number;
    key: string | null;
    optimistic_id?: string | null;
    optimistic_file_path?: string | null;
    optimistic_preview_file_path?: string | null;
    optimistic_imgix_video_file_path?: string | null;
    optimistic_src?: string | null;
    optimistic_preview_src?: string | null;
    optimistic_ready: boolean;
    client_error?: string | null;
};
// ...
export type MiniProject = {
    id: string;
    name: string;
    accessory: string | null;
    private: boolean;
    archived: boolean;
    message_thread_id: string | null;
};
// ...
export type PublicOrganization = {
    id: string;
    avatar_url: string;
    avatar_urls: AvatarUrls;
    name: string;
    slug: string;
    viewer_is_admin: boolean;
    viewer_can_leave: boolean;
};
export type PostLink = {
    id: string;
    name: string;
    url: string;
};
export type Tag = {
    id: string;
    name: string;
    posts_count: number;
    url: string;
    viewer_can_destroy: boolean;
};
export type PollOption = {
    id: string;
    description: string;
    votes_count: number;
    votes_percent: number;
    viewer_voted: boolean;
};
export type Poll = {
    id: string;
    description: string;
    votes_count: number;
    options: PollOption[];
    viewer_voted: boolean;
};
export type FeedbackRequest = {
    id: string;
    has_replied: boolean;
    member: OrganizationMember;
};
export type SubjectFollowUp = {
    id: string;
    member: OrganizationMember;
    show_at: string;
    belongs_to_viewer: boolean;
};
export type ResolvedComment = {
    id: string;
    created_at: string;
    body_html: string;
    url: string;
    viewer_is_author: boolean;
    member: OrganizationMember;
};
export type SyncCustomReaction = {
    id: string;
    name: string;
    file_url: string;
    created_at: string;
};
export type GroupedReaction = {
    viewer_reaction_id: string | null;
    emoji: string | null;
    tooltip: string;
    reactions_count: number;
    custom_content: SyncCustomReaction | null;
};
export type Commenters = {
    latest_commenters: OrganizationMember[];
};
export type PostResolution = {
    resolved_at: string;
    resolved_by: OrganizationMember;
    resolved_html: string | null;
    resolved_comment: ResolvedComment | null;
};
export type ResourceMentionPost = {
    id: string;
    title: string;
    created_at: string;
    published_at: string | null;
    url: string;
};
export type ResourceMentionCall = {
    id: string;
    title: string;
    created_at: string;
    url: string;
};
export type ResourceMentionNote = {
    id: string;
    title: string;
    created_at: string;
    url: string;
};
export type ResourceMention = {
    id: string;
    post: ResourceMentionPost | null;
    call: ResourceMentionCall | null;
    note: ResourceMentionNote | null;
    type_name: string;
};
export type Post = {
    id: string;
    title: string;
    is_title_from_description: boolean;
    created_at: string;
    published_at: string | null;
    published: boolean;
    last_activity_at: string;
    comments_count: number;
    resolved_comments_count: number;
    version: number;
    path: string;
    channel_name: string;
    views_count: number;
    non_member_views_count: number;
    status: 'none' | 'feedback_requested';
    visibility: 'default' | 'public';
    open_graph_image_url: string | null;
    thumbnail_url: string | null;
    unfurled_link: string | null;
    description_html: string;
    truncated_description_html: string;
    is_text_content_truncated: boolean;
    truncated_description_text: string;
    url: string;
    type_name: string;
    organization: PublicOrganization;
    attachments: Attachment[];
    links: PostLink[];
    tags: Tag[];
    poll: Poll | null;
    feedback_requests: FeedbackRequest[] | null;
    follow_ups: SubjectFollowUp[];
    member: OrganizationMember;
    resolved_comment: ResolvedComment | null;
    grouped_reactions: GroupedReaction[];
    project: MiniProject;
    has_parent: boolean;
    has_iterations: boolean;
    viewer_is_organization_member: boolean;
    viewer_is_author: boolean;
    viewer_has_commented: boolean;
    preview_commenters: Commenters;
    viewer_feedback_status: 'none' | 'viewer_requested' | 'open';
    viewer_has_subscribed: boolean;
    viewer_has_viewed: boolean;
    viewer_has_favorited: boolean;
    unseen_comments_count: number;
    viewer_can_resolve: boolean;
    viewer_can_favorite: boolean;
    viewer_can_edit: boolean;
    viewer_can_delete: boolean;
    viewer_can_create_issue: boolean;
    resolution: PostResolution | null;
    latest_comment_preview: string | null;
    latest_comment_path: string | null;
    viewer_is_latest_comment_author: boolean;
    project_pin_id: string | null;
    resource_mentions: ResourceMention[];
};